Plus limit to decimal

+++ src/Extends/MaxNumberIsValidExt.cs
@@ -9,7 +9,7 @@ namespace BitHelp.Core.Validation.Extends
     public static class MaxNumberIsValidExt
     {
         public static ValidationNotification MaxNumberIsValid<TClass>(
-            this ValidationNotification source, TClass data, Expression<Func<TClass, object>> expression, long maximum)
+            this ValidationNotification source, TClass data, Expression<Func<TClass, object>> expression, decimal maximum)
         {
             string prorpety = expression.PropertyTrail();
             object value = expression.Compile().DynamicInvoke(data);
@@ -18,13 +18,13 @@ namespace BitHelp.Core.Validation.Extends
         }
 
         public static ValidationNotification MaxNumberIsValid(
-            this ValidationNotification source, object value, long maximum)
+            this ValidationNotification source, object value, decimal maximum)
         {
             return source.MaxNumberIsValid(value, Resource.DisplayValue, null, maximum);
         }
 
         private static ValidationNotification MaxNumberIsValid(
-            this ValidationNotification source, object value, string display, string reference, long maximum)
+            this ValidationNotification source, object value, string display, string reference, decimal maximum)
         {
             source.LastMessage = null;
             MaxNumberIsValidAttribute validation = new MaxNumberIsValidAttribute(maximum);
